mise a jour decalage + int decalage ->float

+++ includes/rt.h
@@ -79,8 +79,8 @@ typedef struct			s_mat
 	float				diff;
 	float				spec;
 	float				refl;
-	int					offx;
-	int					offy;
+	float				offx;
+	float				offy;
 	SDL_Surface			*txt;
 	SDL_Surface			*txt_bm;
 }						t_mat;
